requirements.txt
seaborn
pandas
tensorflow-gpu
numpy
scikit-learn
matplotlib
scipy